requirements.txt
asyncpg
fastapi
google-genai
langchain
langchain-core
langchain-google-genai
python-dotenv
rapidfuzz